Summary:

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

cc H-Huang awgu wanchaol fegin fduwjj wz337 wconstab d4l3k pragupta msaroufim dcci

[ghstack-poisoned]

+++ test/distributed/tensor/test_utils.py
@@ -54,7 +54,7 @@ class LocalTest(TestCase):
         TP = 8
         mesh_shape = (DP, TP)
         placements = [_StridedShard(0, split_factor=8), Shard(0)]
-        TP_shard_size = int(global_shape[0] / TP)
+        TP_shard_size = global_shape[0] / TP
         for my_coordinate in itertools.product(range(DP), range(TP)):
             local_shape, global_offset = _compute_local_shape_and_global_offset(
                 global_shape, mesh_shape, list(my_coordinate), placements
